+++ src/NexaCRM.Service/Abstractions/Models/TeamModels.cs
@@ -4,7 +4,7 @@ namespace NexaCRM.UI.Models.Teams;
 
 public class TeamInfo
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string TeamCode { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
     public string ManagerName { get; set; } = string.Empty;
@@ -15,8 +15,8 @@ public class TeamInfo
 
 public class TeamMemberInfo
 {
-    public int Id { get; set; }
-    public int TeamId { get; set; }
+    public long Id { get; set; }
+    public long TeamId { get; set; }
     public string TeamName { get; set; } = string.Empty;
     public string Role { get; set; } = string.Empty;
     public string EmployeeCode { get; set; } = string.Empty;
